+++ ChangeLog.md
@@ -9,6 +9,8 @@
 - Updated `README.md` (03/09/2018)
 - Added `UPGRADE.md` (03/09/2018)
 - Added `PageEditFormFactory` + Interface (03/09/2018)
+- Removed SubmitButton from `PageEditType` (03/09/2018)
+
 
 v1.x
 ====